+++ service-api/service-csharp/identity/src/Identity.Application/DependencyInjection.cs
@@ -9,6 +9,7 @@ public static class DependencyInjection
   public static IServiceCollection AddIdentityApplication(this IServiceCollection services)
   {
     services.AddScoped<ListBootstrapTenants>();
+    services.AddScoped<ListBootstrapRoles>();
 
     return services;
   }